backend/requirements.txt
django
djangorestframework
django-cors-headers
beautifulsoup4
langchain-community
lxml
celery
chromadb
sentence-transformers
langchain-chroma
djangorestframework-simplejwt
redis
langchain-huggingface>=0.0.1
channels
channels_redis
aioredis